update transform process to error if the resource doesnt exist and to properly download the relevant files
