+++ .env.example
@@ -1,0 +1,6 @@
+# TRANSFORM VARIABLES
+export COLLECTION=conservation-area
+export DATASET=conservation-area
+export DATASET_NAME=conservation-area
+export TRANSFORMATION_OFFSET=1600
+export TRANSFORMATION_LIMIT=200
