--- src/collection_task.egg-info/PKG-INFO
@@ -1,5 +1,0 @@
-Metadata-Version: 2.4
-Name: collection-task
-Version: 0.1.0
-License-File: LICENSE
-Dynamic: license-file
