--- src/collection_task.egg-info/top_level.txt
@@ -1,1 +1,0 @@
-collection_task
